backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.4.2
inflection==0.5.1
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
gunicorn==21.2.0
django-storages==1.14.2
boto3==1.34.104
python-decouple==3.8
psycopg2-binary==2.9.9